fixed: JoinColumns, remove Enumerate, set CascadeType.ALL, Exception Handlers
